src/feeddb/requirements.txt
Django==1.6.5
South==0.6.2
django-pagination==1.0.5
psycopg2==2.0.13
wsgiref==0.1.2
django-haystack==2.1.0
pysolr==3.2.0
whoosh==2.6.0
